requirements.txt
fastapi
uvicorn
scikit-learn
joblib
numpy
python-jose
passlib[bcrypt]
